% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_authorized}
\alias{check_authorized}
\title{Check whether a response contains status code 401}
\usage{
check_authorized(resp)
}
\arguments{
\item{resp}{The server response as returned by \code{\link[=tasks_run_post]{tasks_run_post()}} or
\code{\link[=tasks_run_get]{tasks_run_get()}}.}
}
\value{
A boolean indicating whether the user is authorized.
}
\description{
Stops with an error message if DATABOARD_SILENT is not TRUE.
}
\seealso{
\code{\link[=da_login]{da_login()}}
}
\keyword{internal}
