% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_succesful}
\alias{check_succesful}
\title{Check whether a response was successful (code 20x)}
\usage{
check_succesful(resp)
}
\arguments{
\item{resp}{The server response as returned by \code{\link[=tasks_run_post]{tasks_run_post()}} or
\code{\link[=tasks_run_get]{tasks_run_get()}}.}
}
\value{
A boolean indicating whether the response was succesful
}
\description{
Stops with an error message if DATABOARD_SILENT is not TRUE.
}
\keyword{internal}
