% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{da_request}
\alias{da_request}
\title{Send a request to the Databoard server}
\usage{
da_request(endpoint, body, wait = 0)
}
\arguments{
\item{endpoint}{The path including a leading slash}

\item{body}{If provided, the body is send via a POST request.
Otherwise, a GET request is issued.}

\item{wait}{Seconds to wait for an answer. Set to 0 to return immediately.}
}
\value{
The response object.
}
\description{
Send a request to the Databoard server
}
