% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genres}
\alias{genres}
\title{Movie genre descriptions}
\format{
\subsection{\code{genres}}{

A data frame with 6 rows and 3 columns:
\describe{
\item{category}{The genre name}
\item{description}{Definition ot the genre.}
\item{example}{Example movies}
}
#' @source Communication Department of the University of Münster (\href{mailto:k.maubach@uni-muenster.de}{k.maubach@uni-muenster.de} and \href{mailto:jakob.juenger@uni-muenster.de}{jakob.juenger@uni-muenster.de}).
}
}
\usage{
genres
}
\description{
A rule set for coding genres
}
\keyword{datasets}
