% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Login-Function for the Databoard API}
\usage{
get_token(username, pw)
}
\arguments{
\item{username}{Username for the Login}

\item{pw}{Password for the Login}
}
\value{
The access Token saved within the system environment
}
\description{
This function logs into the \href{https://databoard.uni-muenster.de/}{Databoard-API} and returns a valid access token. Please contact the \href{https://www.uni-muenster.de/Kowi/en/personen/jakob-juenger.shtml}{Digital Media and Computational Methods} research unit to acquire an user account. Please make sure to keep your login data and access token private.
}
