% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movies}
\alias{movies}
\title{Movie Dataset}
\format{
\subsection{\code{movies}}{

A data frame with 10 rows and 4 columns:
\describe{
\item{id}{A running id number}
\item{name}{Name of the moview}
\item{abstract}{A short abstract, imagined from chatgpt based on the movie title}
\item{year}{A review of the movie, imagined from chatgtp based on the movie title}
}
#' @source Communication Department of the University of Münster (\href{mailto:k.maubach@uni-muenster.de}{k.maubach@uni-muenster.de} and \href{mailto:jakob.juenger@uni-muenster.de}{jakob.juenger@uni-muenster.de}).
}
}
\usage{
movies
}
\description{
A fabricated data set containing information about different movies
}
\keyword{datasets}
