% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_json}
\alias{parse_json}
\title{Parse JSON from a response}
\usage{
parse_json(resp)
}
\arguments{
\item{resp}{The server response as returned by \code{\link[=tasks_run_post]{tasks_run_post()}} or
\code{\link[=tasks_run_get]{tasks_run_get()}}.}
}
\value{
The parsed JSON object or NULL
}
\description{
Only try to parse JSON if the server actually returned JSON.
For example, on HTTP errors the body may be an HTML error page.
}
\keyword{internal}
