% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{refi_guid}
\alias{refi_guid}
\title{Generate a GUID for REFI-QDA objects}
\usage{
refi_guid()
}
\value{
A single upper-case UUID string.
}
\description{
Generate a GUID for REFI-QDA objects
}
\keyword{internal}
