% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{refi_timestamp}
\alias{refi_timestamp}
\title{Current time as REFI-QDA timestamp}
\usage{
refi_timestamp()
}
\value{
A single character value in ISO 8601 format (UTC), e.g.
\code{"2024-05-01T12:00:00Z"}.
}
\description{
Current time as REFI-QDA timestamp
}
\keyword{internal}
