% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Automated content coding with llms}
\usage{
run_pipeline(input, access_token, system_prompt, temperature = 1)
}
\arguments{
\item{input}{Text for coding, ideally a dataframe with an identifier (e.g. ID) and the text for coding.}

\item{access_token}{Your personal access token to the databoard service. For generating use the get_token function.}

\item{system_prompt}{The coding rules for your specific project. Should be in the form of a txt.file in the same folder. See \href{https://databoard.uni-muenster.de/}{here} for examples.}

\item{temperature}{The desired temperature (e.g. the level of answer-randomness). Add numerical value, the default is set to 1.}
}
\value{
Your original dataframe with additional columns with the coded results.
}
\description{
This function uses the databoard service from the \href{https://www.uni-muenster.de/Kowi/en/personen/jakob-juenger.shtml}{Digital Media and Computational Methods} research unit to facilitate automated content coding with llms.
}
\examples{
input <- "This is a comment about love"
access_token <- "your personal access token"
system_prompt <- "Code when the word 'love' is mentioned"
run_pipeline(input, access_token, system_prompt)
}
