% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_task.R
\name{submit_task}
\alias{submit_task}
\title{Automated content coding with llms}
\usage{
submit_task(input, access_token, system_prompt)
}
\arguments{
\item{input}{Text for coding, ideally a dataframe with an identifier (e.g. ID) and the text for coding.}

\item{access_token}{Your personal access token to the databoard service. For generating use the get_token function.}

\item{system_prompt}{The coding rules for your specific project. Should be in the form of a txt.file in the same folder. See \href{https://databoard.uni-muenster.de/}{here} for examples.}
}
\value{
Your original dataframe with additional columns with the coded results.
}
\description{
Automated content coding with llms
}
\examples{
input <- "This is a comment about love"
access_token <- "your personal access token"
system_prompt <- "Code when the word 'love' is mentioned"
run_pipeline(input, access_token, system_prompt)
}
